+++ ChangeLog.md
@@ -27,6 +27,10 @@ See docs/process.md for more on how version tagging works.
 - The internal `read_` and `readAsync` functions no longer handle date URIs.
   This only effects builds that use `-sSINGLE_FILE` or `--memory-init-file`.
   (#19792)
+- The `asm` property of the Module object (which held the raw exports of the
+  wasm module) has been removed.  Internally, this is now accessed via the
+  `wasmExports` global. If necessary, it is possible to export `wasmExports`
+  on the Module object using `-sEXPORTED_RUNTIME_METHODS=wasmExports`. (#19816)
 
 3.1.43 - 07/10/23
 -----------------